Adds Sass
I added Sass to the project and started the process of moving over from css to scss. I'm trying to implement a 7-1 structure and using a better CSS reset.

In hindsight I should have done this from the start - but this is a good time to review and use the things I've been learning the last month.