Updates Navbar (again!)
I made the A.Berg tag on the anchor also link back to the Home page... it just made sense for it to do that.

I also refactored the structure of the title in the Nav. Instead of a free-floating h2, it's now an h2 wrapped in a Link and another div. This just helped me to better conceptualize the nav as a set of containers. I may have to place the NavLinks in their own divs as well - so I can style those later on. Right now they're very basic.