Fixes confetti
The positioning problem with the confetti has been solved! I put the Rewards component and the button element in another div, and centered them using text-align.
I didn't want to create a whole new class for this, but if I examine my utility partial I might find it already exists 🤔🤔🤔